Playlist task

+++ src/main/java/ru/tinkoff/tinkoffmusicplatform/repository/SongRepository.java
@@ -5,7 +5,6 @@ import org.springframework.stereotype.Repository;
 import ru.tinkoff.tinkoffmusicplatform.data.Song;
 
 @Repository
-public interface SongRepository extends CrudRepository<Song, Integer> {
-
+public interface SongRepository extends CrudRepository<Song, Long> {
 
 }
